# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/api/admin/posts": {
            "post": {
                "tags": [
                    "Admin Posts"
                ],
                "summary": "ایجاد پست جدید توسط ادمین",
                "operationId": "50f2470f1c348fad74012869a1ff70a5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "content",
                                    "status"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "عنوان پست"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "متن کامل پست"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "draft",
                                            "published",
                                            "archived"
                                        ]
                                    },
                                    "image": {
                                        "description": "آپلود تصویر (اختیاری)",
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "پست با موفقیت ایجاد شد",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "پست با موفقیت ایجاد شد"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "خطای اعتبارسنجی"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/my-posts": {
            "get": {
                "tags": [
                    "Admin Posts"
                ],
                "summary": "دریافت پست‌های ادمین جاری",
                "operationId": "42d51b5ae6a9327d57742abf884a999b",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "عبارت جستجو در عنوان پست",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "لیست پست‌ها",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Post"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/posts/{id}": {
            "post": {
                "tags": [
                    "Admin Posts"
                ],
                "summary": "ویرایش پست",
                "operationId": "df52ffa8a33b3b4da6de0ea41588b62c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "شناسه پست",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "عنوان جدید"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "محتوای جدید"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "draft",
                                            "published",
                                            "archived"
                                        ]
                                    },
                                    "image": {
                                        "description": "تصویر جدید (اختیاری)",
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "پست با موفقیت ویرایش شد",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "پست با موفقیت به‌روزرسانی شد"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "دسترسی غیرمجاز"
                    },
                    "404": {
                        "description": "پست یافت نشد"
                    },
                    "422": {
                        "description": "خطای اعتبارسنجی"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin Posts"
                ],
                "summary": "حذف پست",
                "operationId": "766270a7c7c0f7d4d0278d10da208e26",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "شناسه پست",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "پست حذف شد",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "پست با موفقیت حذف شد"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "دسترسی غیرمجاز"
                    },
                    "404": {
                        "description": "پست یافت نشد"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}